refactor for rest with new scheduler types
