+++ core/plugin.go
@@ -1,0 +1,6 @@
+package core
+
+type Plugin interface {
+	Name() string
+	Version() int
+}
