+++ core/task.go
@@ -18,6 +18,22 @@ type Task interface {
 	MissedCount() uint
 	LastRunTime() time.Time
 	CreationTime() time.Time
+	DeadlineDuration() time.Duration
+	SetDeadlineDuration(time.Duration)
+	Option(...TaskOption) TaskOption
+}
+
+type TaskOption func(Task) TaskOption
+
+// TaskDeadlineDuration sets the tasks deadline.
+// The deadline is the amount of time that can pass before a worker begins
+// processing the tasks collect job.
+func TaskDeadlineDuration(v time.Duration) TaskOption {
+	return func(t Task) TaskOption {
+		previous := t.DeadlineDuration()
+		t.SetDeadlineDuration(v)
+		return TaskDeadlineDuration(previous)
+	}
 }
 
 type TaskErrors interface {
