+++ core/workflow.go
@@ -12,4 +12,24 @@ type Workflow interface {
 	State() WorkflowState
 }
 
-type WfMap struct{}
+type WfMap struct {
+	Collect CollectStep
+	Process ProcessStep
+	Publish PublishStep
+}
+
+type CollectStep struct {
+	MetricTypes []MetricType
+	Process     []ProcessStep
+	Publish     []PublishStep
+}
+
+type ProcessStep struct {
+	Plugin  Plugin
+	Process []ProcessStep
+	Publish []PublishStep
+}
+
+type PublishStep struct {
+	Plugin Plugin
+}
